+++ pulsar/utils/structures.py
@@ -161,6 +161,12 @@ class AttributeDictionary(object):
     def __setattr__(self, name, value):
         self.__dict__[name] = value
         
+    def __setitem__(self, name, value):
+        self.__dict__[name] = value
+        
+    def __getitem__(self, name):
+        return self.__dict__[name]
+        
     def update(self, iterable):
         for name, value in mapping_iterator(iterable):
             setattr(self, name, value)
@@ -174,7 +180,7 @@ class AttributeDictionary(object):
     
 def merge_prefix(deque, size):
     """Replace the first entries in a deque of bytes with a single
-string of up to size bytes."""
+string of up to *size* bytes."""
     if len(deque) == 1 and len(deque[0]) <= size:
         return
     prefix = []