+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
-Version 0.4 - 2012-Oct-23
+Version 0.4 - Development
 ============================
 * Overall refactoring of API and therefore incompatible with previous versions.
+* Development status set to ``Beta``.
 * Support pypy_ and python 3.3.
 * Added the new :mod:`pulsar.utils.httpurl` module for HTTP tools and HTTP clients.
 * Refactored :class:`pulsar.Deferred` to be more compatible with twisted. You
@@ -13,7 +14,7 @@ Version 0.4 - 2012-Oct-23
   a dictionary.
 * Socket and WSGI :class:`pulsar.Application` are built on top of the new
   :class:`pulsar.AsyncSocketServer` framework class.
-* **218 regression tests**, **79% coverage**.
+* **267 regression tests**, **79% coverage**.
 
 Version 0.3 - 2012-May-03
 ============================